+++ packages/global/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.1](https://github.com/bolt-design-system/bolt/tree/master/packages/global/compare/v2.12.0...v2.12.1) (2019-12-12)
+
+**Note:** Version bump only for package @bolt/global
+
+
+
+
+
+# [2.12.0](https://github.com/bolt-design-system/bolt/tree/master/packages/global/compare/v2.11.4...v2.12.0) (2019-11-26)
+
+
+### Bug Fixes
+
+* properly deprecate old figure media props, update old demos ([0022b17](https://github.com/bolt-design-system/bolt/tree/master/packages/global/commit/0022b17))
+
+
+
+
+
+## [2.11.2](https://github.com/bolt-design-system/bolt/tree/master/packages/global/compare/v2.11.1...v2.11.2) (2019-11-14)
+
+**Note:** Version bump only for package @bolt/global
+
+
+
+
+
+# [2.11.0](https://github.com/bolt-design-system/bolt/tree/master/packages/global/compare/v2.10.0...v2.11.0) (2019-11-14)
+
+**Note:** Version bump only for package @bolt/global
+
+
+
+
+
 ## [2.10.1](https://github.com/bolt-design-system/bolt/tree/master/packages/global/compare/v2.10.0...v2.10.1) (2019-11-05)
 
 **Note:** Version bump only for package @bolt/global